test(ProductServiceTest): 상품 조회 테스트

+++ src/main/java/woowacourse/shoppingcart/product/application/dto/response/ProductResponse.java
@@ -6,13 +6,13 @@ public class ProductResponse {
 
     private long id;
     private String name;
-    private int price;
+    private long price;
     private String image;
 
     public ProductResponse() {
     }
 
-    public ProductResponse(final long id, final String name, final int price, final String image) {
+    public ProductResponse(final long id, final String name, final long price, final String image) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -31,7 +31,7 @@ public class ProductResponse {
         return name;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
